Drop Python 2.7 support, Switch to GitHub Actions (#116)
This officially drops support for Python 2, making the new requirement
Python >= 3.6.

In this process, setup.py / setup.cfg was dropped in favor of
pyproject.toml with poetry as the build-system tool.

This also switches away from Travis-CI to GitHub Actions due to the
impending changes that will affect open-source projects.

In the process of dropping support for Python 2, several bugs that were
causing issues in Python 3.6+ were fixed:

- The number of points in numpy.linspace was explicitly made to be
`int`, but previously this was implicit
- numpy.squeeze was changed, and the angles array had to be changed to
fix it
- the currat-axe spurion calculator test was bugged and adjusted

No new features or major bug fixes are introduced with this version,
this is merely to drop support for Python 2, and to ensure all tests are
passing in Python 3.6+.

+++ neutronpy/crystal/structure_factors.py
@@ -123,9 +123,9 @@ class MagneticFormFactor(object):
 
         if q is None:
             if qrange is None:
-                q = np.linspace(0., 2., 2. / 0.025 + 1)
+                q = np.linspace(0., 2., int(2. / 0.025) + 1)
             else:
-                q = np.linspace(qrange[0], qrange[1], (qrange[1] - qrange[0]) / 0.025 + 1)
+                q = np.linspace(qrange[0], qrange[1], int((qrange[1] - qrange[0]) / 0.025) + 1)
         if g is None:
             g = 2.
